requirements.txt
python-dotenv==1.0.0
pydantic==1.10.7
aiofiles==23.2.1
aiohttp==3.8.5
typer[all]==0.9.0
tabulate==0.9.0
PyYAML==6.0.1
aiohttp-pydantic==1.12.2